# Conflicts:
#	core/network/src/commonMain/kotlin/com/project/network/product_network/ProductApiClient.kt
#	features/application-sections/chat-section/chats-impl/src/commonMain/kotlin/com/project/chats/screens/chats/viewmodel/ChatsViewModel.kt

+++ features/application-sections/profile-section/warehouse-feature/warehouse-impl/src/commonMain/kotlin/com/profile/profile/screens/main_refactor/screens/arrival_and_consumption_goods/model/ProductArrivalAndConsumption.kt
@@ -4,5 +4,7 @@ data class ProductArrivalAndConsumption (
 
     val product: AllProductArrivalAndConsumption,   // ID товара
 
-    var count: Int // Количество товаров
+    var count: Double // Количество товаров
 )
+
+
